requirements.txt
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
schedule>=1.2.0
python-dateutil>=2.8.2
groq>=0.9.0
feedparser>=6.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pytest>=7.4.0
libsql-experimental>=0.0.1
